only search within resources if a query string was set
In initial implementation, the search would always search for resources
with an empty string, and come up with a list of models that had any
resources in them. This list would then be added as a "must" expression
to opensearch, limiting the main search too much.

Now matching resources will only be searched if a query string was set.